[MOV] im_livechat: move Thread outside of public_livechat.js
Task-2212347
