+++ addons/im_livechat/__manifest__.py
@@ -60,6 +60,7 @@ Help your customers with this chat, and analyse their feedback.
         'web.qunit_suite_tests': [
             'im_livechat/static/tests/qunit_suite_tests/components/**/*.js',
             'im_livechat/static/src/legacy/public_livechat.js',
+            'im_livechat/static/src/legacy/widgets/*',
         ],
         'web.assets_qweb': [
             'im_livechat/static/src/components/*/*.xml',
@@ -158,6 +159,7 @@ Help your customers with this chat, and analyse their feedback.
             'bus/static/src/js/services/bus_service.js',
             'mail/static/src/js/utils.js',
             'im_livechat/static/src/legacy/public_livechat.js',
+            'im_livechat/static/src/legacy/widgets/*',
 
             ('include', 'web._assets_helpers'),
